Reduce duplication of code within the `analog::adc` module
