+++ esp-hal/src/analog/adc/mod.rs
@@ -21,6 +21,8 @@
 //! }
 //! ```
 
+use core::marker::PhantomData;
+
 pub use self::implementation::*;
 
 #[cfg_attr(esp32, path = "esp32.rs")]
@@ -53,6 +55,95 @@ pub enum Attenuation {
 pub enum AdcCalSource {
     Gnd,
     Ref,
+}
+
+/// An I/O pin which can be read using the ADC.
+pub struct AdcPin<PIN, ADCI, CS = ()> {
+    pub pin: PIN,
+    #[cfg_attr(esp32, allow(unused))]
+    pub cal_scheme: CS,
+    _phantom: PhantomData<ADCI>,
+}
+
+#[cfg(feature = "embedded-hal-02")]
+impl<PIN, ADCI, CS> embedded_hal_02::adc::Channel<ADCI> for AdcPin<PIN, ADCI, CS>
+where
+    PIN: embedded_hal_02::adc::Channel<ADCI, ID = u8>,
+{
+    type ID = u8;
+
+    fn channel() -> Self::ID {
+        PIN::channel()
+    }
+}
+
+/// Configuration for the ADC.
+pub struct AdcConfig<ADCI> {
+    pub resolution: Resolution,
+    pub attenuations: [Option<Attenuation>; NUM_ATTENS],
+    _phantom: PhantomData<ADCI>,
+}
+
+impl<ADCI> AdcConfig<ADCI> {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn enable_pin<PIN>(&mut self, pin: PIN, attenuation: Attenuation) -> AdcPin<PIN, ADCI>
+    where
+        PIN: AdcChannel,
+    {
+        self.attenuations[PIN::CHANNEL as usize] = Some(attenuation);
+
+        AdcPin {
+            pin,
+            cal_scheme: AdcCalScheme::<()>::new_cal(attenuation),
+            _phantom: PhantomData,
+        }
+    }
+
+    #[cfg(not(esp32))]
+    pub fn enable_pin_with_cal<PIN, CS>(
+        &mut self,
+        pin: PIN,
+        attenuation: Attenuation,
+    ) -> AdcPin<PIN, ADCI, CS>
+    where
+        ADCI: CalibrationAccess,
+        PIN: AdcChannel,
+        CS: AdcCalScheme<ADCI>,
+    {
+        self.attenuations[PIN::CHANNEL as usize] = Some(attenuation);
+
+        AdcPin {
+            pin,
+            cal_scheme: CS::new_cal(attenuation),
+            _phantom: PhantomData,
+        }
+    }
+}
+
+impl<ADCI> Default for AdcConfig<ADCI> {
+    fn default() -> Self {
+        Self {
+            resolution: Resolution::default(),
+            attenuations: [None; NUM_ATTENS],
+            _phantom: PhantomData,
+        }
+    }
+}
+
+#[cfg(not(esp32))]
+#[doc(hidden)]
+pub trait CalibrationAccess: RegisterAccess {
+    const ADC_CAL_CNT_MAX: u16;
+    const ADC_CAL_CHANNEL: u16;
+    const ADC_VAL_MASK: u16;
+
+    fn enable_vdef(enable: bool);
+
+    /// Enable internal calibration voltage source
+    fn connect_cal(source: AdcCalSource, enable: bool);
 }
 
 /// A helper trait to get the ADC channel of a compatible GPIO pin.
@@ -105,3 +196,24 @@ trait AdcCalEfuse {
     /// Returns digital value for reference voltage for a given attenuation
     fn get_cal_code(atten: Attenuation) -> Option<u16>;
 }
+
+macro_rules! impl_adc_interface {
+    ($adc:ident [
+        $( ($pin:ident, $channel:expr) ,)+
+    ]) => {
+        $(
+            impl $crate::analog::adc::AdcChannel for crate::gpio::$pin<crate::gpio::Analog> {
+                const CHANNEL: u8 = $channel;
+            }
+
+            #[cfg(feature = "embedded-hal-02")]
+            impl embedded_hal_02::adc::Channel<crate::peripherals::$adc> for crate::gpio::$pin<crate::gpio::Analog> {
+                type ID = u8;
+
+                fn channel() -> u8 { $channel }
+            }
+        )+
+    }
+}
+
+pub(crate) use impl_adc_interface;
