+++ esp-hal/src/analog/adc/xtensa.rs
@@ -1,10 +1,6 @@
-use core::marker::PhantomData;
-
 #[cfg(esp32s3)]
 pub use self::calibration::*;
-#[cfg(esp32s3)]
-use super::AdcCalEfuse;
-use super::{AdcCalScheme, AdcCalSource, AdcChannel, Attenuation};
+use super::{AdcCalScheme, AdcCalSource, AdcChannel, AdcConfig, AdcPin, Attenuation};
 #[cfg(esp32s3)]
 use crate::efuse::Efuse;
 use crate::{
@@ -13,6 +9,8 @@ use crate::{
 };
 
 mod calibration;
+
+pub(super) const NUM_ATTENS: usize = 10;
 
 // Constants taken from:
 // https://github.com/espressif/esp-idf/blob/903af13e8/components/soc/esp32s2/include/soc/regi2c_saradc.h
@@ -74,81 +72,20 @@ cfg_if::cfg_if! {
 }
 
 /// The sampling/readout resolution of the ADC.
-#[derive(PartialEq, Eq, Clone, Copy)]
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
 pub enum Resolution {
+    #[default]
     Resolution13Bit,
-}
-
-/// An I/O pin which can be read using the ADC.
-pub struct AdcPin<PIN, ADCI, CS = ()> {
-    pub pin: PIN,
-    pub cal_scheme: CS,
-    _phantom: PhantomData<ADCI>,
-}
-
-#[cfg(feature = "embedded-hal-02")]
-impl<PIN, ADCI, CS> embedded_hal_02::adc::Channel<ADCI> for AdcPin<PIN, ADCI, CS>
-where
-    PIN: embedded_hal_02::adc::Channel<ADCI, ID = u8>,
-{
-    type ID = u8;
-
-    fn channel() -> Self::ID {
-        PIN::channel()
-    }
-}
-
-/// Configuration for the ADC.
-pub struct AdcConfig<ADCI> {
-    pub resolution: Resolution,
-    pub attenuations: [Option<Attenuation>; 10],
-    _phantom: PhantomData<ADCI>,
 }
 
 impl<ADCI> AdcConfig<ADCI>
 where
     ADCI: RegisterAccess,
 {
-    pub fn new() -> AdcConfig<ADCI> {
-        Self::default()
-    }
-
-    pub fn enable_pin<PIN>(&mut self, pin: PIN, attenuation: Attenuation) -> AdcPin<PIN, ADCI>
-    where
-        PIN: AdcChannel,
-    {
-        self.attenuations[PIN::CHANNEL as usize] = Some(attenuation);
-
-        AdcPin {
-            pin,
-            cal_scheme: AdcCalScheme::<()>::new_cal(attenuation),
-            _phantom: PhantomData,
-        }
-    }
-
-    pub fn enable_pin_with_cal<PIN, CS>(
-        &mut self,
-        pin: PIN,
-        attenuation: Attenuation,
-    ) -> AdcPin<PIN, ADCI, CS>
-    where
-        ADCI: CalibrationAccess,
-        PIN: AdcChannel,
-        CS: AdcCalScheme<ADCI>,
-    {
-        self.attenuations[PIN::CHANNEL as usize] = Some(attenuation);
-
-        AdcPin {
-            pin,
-            cal_scheme: CS::new_cal(attenuation),
-            _phantom: PhantomData,
-        }
-    }
-
     /// Calibrate ADC with specified attenuation and voltage source
     pub fn adc_calibrate(atten: Attenuation, source: AdcCalSource) -> u16
     where
-        ADCI: CalibrationAccess,
+        ADCI: super::CalibrationAccess,
     {
         let mut adc_max: u16 = 0;
         let mut adc_min: u16 = u16::MAX;
@@ -191,16 +128,6 @@ where
     }
 }
 
-impl<ADCI> Default for AdcConfig<ADCI> {
-    fn default() -> Self {
-        AdcConfig {
-            resolution: Resolution::Resolution13Bit,
-            attenuations: [None; 10],
-            _phantom: PhantomData,
-        }
-    }
-}
-
 #[doc(hidden)]
 pub trait RegisterAccess {
     fn adc_samplecfg(channel: u16);
@@ -230,18 +157,6 @@ pub trait RegisterAccess {
 
     /// Reset flags
     fn reset();
-}
-
-#[doc(hidden)]
-pub trait CalibrationAccess: RegisterAccess {
-    const ADC_CAL_CNT_MAX: u16;
-    const ADC_CAL_CHANNEL: u16;
-    const ADC_VAL_MASK: u16;
-
-    fn enable_vdef(enable: bool);
-
-    /// Enable internal calibration voltage source
-    fn connect_cal(source: AdcCalSource, enable: bool);
 }
 
 impl RegisterAccess for crate::peripherals::ADC1 {
@@ -361,7 +276,7 @@ impl RegisterAccess for crate::peripherals::ADC1 {
 }
 
 #[cfg(esp32s3)]
-impl CalibrationAccess for crate::peripherals::ADC1 {
+impl super::CalibrationAccess for crate::peripherals::ADC1 {
     const ADC_CAL_CNT_MAX: u16 = ADC_CAL_CNT_MAX;
     const ADC_CAL_CHANNEL: u16 = ADC_CAL_CHANNEL;
     const ADC_VAL_MASK: u16 = ADC_VAL_MASK;
@@ -500,7 +415,7 @@ impl RegisterAccess for crate::peripherals::ADC2 {
 }
 
 #[cfg(esp32s3)]
-impl CalibrationAccess for crate::peripherals::ADC2 {
+impl super::CalibrationAccess for crate::peripherals::ADC2 {
     const ADC_CAL_CNT_MAX: u16 = ADC_CAL_CNT_MAX;
     const ADC_CAL_CHANNEL: u16 = ADC_CAL_CHANNEL;
     const ADC_VAL_MASK: u16 = ADC_VAL_MASK;
@@ -649,7 +564,7 @@ where
 }
 
 #[cfg(esp32s3)]
-impl AdcCalEfuse for crate::peripherals::ADC1 {
+impl super::AdcCalEfuse for crate::peripherals::ADC1 {
     fn get_init_code(atten: Attenuation) -> Option<u16> {
         Efuse::get_rtc_calib_init_code(1, atten)
     }
@@ -664,7 +579,7 @@ impl AdcCalEfuse for crate::peripherals::ADC1 {
 }
 
 #[cfg(esp32s3)]
-impl AdcCalEfuse for crate::peripherals::ADC2 {
+impl super::AdcCalEfuse for crate::peripherals::ADC2 {
     fn get_init_code(atten: Attenuation) -> Option<u16> {
         Efuse::get_rtc_calib_init_code(2, atten)
     }
@@ -725,31 +640,8 @@ where
     }
 }
 
-macro_rules! impl_adc_interface {
-    ($adc:ident [
-        $( ($pin:ident, $channel:expr) ,)+
-    ]) => {
-
-        $(
-            impl $crate::analog::adc::AdcChannel for crate::gpio::$pin<crate::gpio::Analog> {
-                const CHANNEL: u8 = $channel;
-            }
-
-            #[cfg(feature = "embedded-hal-02")]
-            impl embedded_hal_02::adc::Channel<$adc> for crate::gpio::$pin<crate::gpio::Analog> {
-                type ID = u8;
-
-                fn channel() -> u8 { $channel }
-            }
-        )+
-    }
-}
-
 mod adc_implementation {
-    #[cfg(feature = "embedded-hal-02")]
-    use crate::peripherals::{ADC1, ADC2};
-
-    impl_adc_interface! {
+    crate::analog::adc::impl_adc_interface! {
         ADC1 [
             (Gpio1,  0),
             (Gpio2,  1),
@@ -764,7 +656,7 @@ mod adc_implementation {
         ]
     }
 
-    impl_adc_interface! {
+    crate::analog::adc::impl_adc_interface! {
         ADC2 [
             (Gpio11, 0),
             (Gpio12, 1),
